Added inspector implementation of 3D graphs visuals
