+++ src/anemoi/graphs/plotting/style.py
@@ -1,0 +1,3 @@
+
+annotations_style = {"text": "", "showarrow": False, "xref": "paper", "yref": "paper", "x": 0.005, "y": -0.002}
+plotly_axis_config = {"showgrid": False, "zeroline": False, "showticklabels": False}
